fix(package.json): :bulb: 更新vite版本
