+++ example/vite.config.ts
@@ -1,17 +1,19 @@
 import { UserConfig } from "vite"
 import { resolve } from "path"
 import vueMd from "../src/index"
+import vuePlugins from "@vitejs/plugin-vue"
 
 function pathResolve(dir: string) {
   return resolve(__dirname, dir)
 }
+
 const viteConfig: UserConfig = {
-  port: 3050,
   alias: {
-    "/@/": pathResolve("./src/"),
+    "@": pathResolve("./src/"),
   },
   plugins: [
     vueMd(),
+    vuePlugins(),
   ],
   optimizeDeps: {
     include: ["element-plus", "highlight.js"],
