+++ package.json
@@ -33,35 +33,36 @@
     }
   },
   "dependencies": {
-    "@vue/compiler-sfc": "^3.0.4",
+    "@vue/compiler-sfc": "^3.0.5",
     "fs-extra": "^9.0.1",
     "highlight.js": "^10.4.1",
     "markdown-it": "^12.0.3",
     "markdown-it-container": "^3.0.0",
-    "vite": "^1.0.0-rc.13"
+    "vite": "^2.0.0-beta.5"
   },
   "peerDependencies": {
-    "vite": ">=1.0.0-rc.13"
+    "vite": ">=2.0.0-beta.5"
   },
   "devDependencies": {
+    "@commitlint/cli": "^11.0.0",
+    "@commitlint/config-conventional": "^11.0.0",
     "@types/fs-extra": "^9.0.5",
     "@types/markdown-it": "^12.0.0",
     "@types/markdown-it-container": "^2.0.3",
     "@types/typescript": "^2.0.0",
     "@typescript-eslint/eslint-plugin": "^4.6.1",
     "@typescript-eslint/parser": "^4.6.1",
-    "@commitlint/cli": "^11.0.0",
-    "@commitlint/config-conventional": "^11.0.0",
+    "@vitejs/plugin-vue": "^1.0.4",
+    "commitizen": "^4.2.2",
     "conventional-changelog-cli": "^2.1.1",
     "cz-conventional-changelog": "^3.3.0",
-    "commitizen": "^4.2.2",
-    "husky": "^4.3.0",
-    "typescript": "^4.1.2",
     "eslint": "^7.12.1",
     "eslint-config-prettier": "^6.15.0",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-prettier": "^3.1.4",
     "eslint-plugin-vue": "^7.1.0",
-    "lint-staged": "^10.5.1"
+    "husky": "^4.3.0",
+    "lint-staged": "^10.5.1",
+    "typescript": "^4.1.2"
   }
 }
